requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
python-dotenv==1.0.1
groq==0.4.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.26.0 
motor